Move ThreadPost and TimelinePost to :ui module, because these are view state not business data.
